requirements.txt
discord.py
requests
peewee
python-dotenv
peewee
google-auth-oauthlib 
python-gemini-api 
google-api-python-client 
google-auth-httplib2
pillow
